# swagger.yaml
swagger: "2.0"
info:
  description: "API documentation for streaming service"
  version: "1.0.0"
  title: "Streaming Service API"
host: "localhost:5000"
schemes:
  - "http"
paths:
  /start-stream:
    post:
      summary: "Start a new stream"
      description: "This endpoint starts a new stream and returns the stream details."
      responses:
        200:
          description: "Stream started successfully"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
              stream_data:
                type: "object"
                properties:
                  stream_uuid:
                    type: "string"
                  id:
                    type: "string"
                  name:
                    type: "string"
                  public:
                    type: "boolean"
                  record:
                    type: "boolean"
        500:
          description: "Failed to start stream"
  /stop-stream/{stream_id}:
    post:
      summary: "Stop a stream"
      description: "This endpoint stops a stream by its ID."
      parameters:
        - name: "stream_id"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "Stream stopped successfully"
        500:
          description: "Failed to stop stream"
  /list-streams:
    get:
      summary: "List all streams"
      description: "This endpoint lists all active streams."
      responses:
        200:
          description: "List of streams"
          schema:
            type: "array"
            items:
              type: "object"
              properties:
                id:
                  type: "string"
                name:
                  type: "string"
                public:
                  type: "boolean"
                record:
                  type: "boolean"
        500:
          description: "Failed to list streams"
  /socket.io:
    get:
      summary: "WebSocket endpoint"
      description: "WebSocket endpoint for real-time communication"
      responses:
        101:
          description: "Switching Protocols"